Define modules as paths to functions, not functions
This is to avoid conflicts when imported from multiple locations.

Example: Module nur.repos.foo.modules.bar defines an option.
If nur.repos.foo.modules.bar is a function, importing it from
multiple places will result in defining the option multiple times,
which the module system doesn't support.

If it instead is a path, there won't be a conflict, the option
will only be defined once.