repos.ma27: nur-packages -> nixexprs
As I renamed the repository in GitLab it's still possible to clone
`nur-packages`, however it's IMHO better to stay in sync with the
upstream settings.

The rename happened as `-packages` is an unsuitable suffix as I use the
repo for modules and library functions as well.

src: f63620dd475617b7b3adef8d4e11da93bfa63aef