added URL to the list of arguments
There might be an issue in CI where the nur-combined directory does not
get cloned. To remedy this, we pass the url of a repos.json element to
the indexing function, with as goal resolving the "pkgs" prefix, as it
seemed to come up quite often. There are some other prefixes that I
haven't figured out yet, but at least this is a start