Tried to fix a bunch of missing attr paths
I'm not sure if this is genuinely the case, but fro the CI logs it seems
to me that nur-combined does not actually get used when evaluating
repositories, which was an assumption that I previously had. The current
addition tries it's best to resolve the issue of default.nix not
refering to a local path, but to a derivation in the nix store.
Currently, only gitlab and github links are supported, as smaller git
hosts do not get special treatment from nixpkgs, and thus their
derivation name is not "source"